LPD-88081 Move per-layout Style Book resolver to StyleBookEntryProviderUtil
